Refactor 'l1script' and 'l2script' entities into its own file
